requirements.txt
Django==4.2.1
djangorestframework==3.14.0
psycopg==3.1.9
psycopg-binary==3.1.9
django-enumchoicefield==3.0.1